Slowly moving things over to a cleaner architecture
