+++ Guppi.Domain/Guppi.Domain.csproj
@@ -1,12 +1,11 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netstandard2.0</TargetFramework>
+    <TargetFramework>net5.0</TargetFramework>
   </PropertyGroup>
 
   <ItemGroup>
     <Folder Include="Entities\" />
-    <Folder Include="Common\" />
     <Folder Include="Interfaces\" />
   </ItemGroup>
